+++ packages/backend/src/nest/storage/channels/channels.service.spec.ts
@@ -1226,8 +1226,34 @@ describe('ChannelsService', () => {
       )
     })
 
-    it('accepts channel metadata deletion from a sigchain admin', async () => {
-      await expectChannelEntryValidation(channelDelEntry('channel-id-to-delete'), aliceUserId, true)
+    it('accepts public channel metadata deletion from a sigchain admin', async () => {
+      const publicChannel = await factory.build<PublicChannel>('PublicChannel', {
+        owner: aliceUserId,
+        teamId: community.teamId!,
+        public: true,
+      })
+      await channelsService.setChannel(publicChannel)
+      await expectChannelEntryValidation(channelDelEntry(publicChannel.id), aliceUserId, true, 'public')
+    })
+
+    it('accepts public channel metadata deletion from a sigchain admin even if no channel is found', async () => {
+      await expectChannelEntryValidation(channelDelEntry('this-is-a-random-channel-id'), aliceUserId, true, 'public')
+    })
+
+    it('accepts private channel metadata deletion from a sigchain admin with found channel', async () => {
+      const channelRoleName = sigChainService.activeChain.channels.create()
+      const privateChannel = await factory.build<PublicChannel>('PublicChannel', {
+        owner: aliceUserId,
+        teamId: community.teamId!,
+        public: false,
+        roleName: channelRoleName,
+      })
+      await channelsService.setChannel(privateChannel)
+      await expectChannelEntryValidation(channelDelEntry(privateChannel.id), aliceUserId, true, 'private')
+    })
+
+    it('rejects private channel metadata deletion from a sigchain admin when no channel is found', async () => {
+      await expectChannelEntryValidation(channelDelEntry('this-is-a-random-channel-id'), aliceUserId, false, 'private')
     })
 
     it('rejects channel metadata deletion from a non-admin member', async () => {