+++ packages/backend/src/nest/storage/orbitDb/orbitDb.service.ts
@@ -217,20 +217,6 @@ export class OrbitDbService {
       }) as any
     )
     orbitDbUseAccessController(
-      this.channelMetadataAccessController.createAccessControllerFunc({
-        write: ['*'],
-        sigchainService: this.sigChainService,
-        isPublic: true,
-      }) as any
-    )
-    orbitDbUseAccessController(
-      this.channelMetadataAccessController.createAccessControllerFunc({
-        write: ['*'],
-        sigchainService: this.sigChainService,
-        isPublic: false,
-      }) as any
-    )
-    orbitDbUseAccessController(
       this.userProfileAccessController.createAccessControllerFunc({
         write: ['*'],
         sigchainService: this.sigChainService,